requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-modeltranslation==0.18.9
idna==3.4
oauthlib==3.2.2
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
